extract UpdateConversationReceiptMode
